project_folder/requirements.txt
python-dotenv
requests
msal
rich
paramiko
httpx
bcrypt
pathspec